Add: source files
Source for this project were added which include metadata.py, reference.py,
item.py and block.py. 'metadata' refers to data associated with reference while
'reference' is object that wraps another python object associating it with priority.

'priority' is number to be used when ranking and sorting references. 'item' is object
that wraps reference and associate it with priority and matadata. 'block' is object
that wraps multiple item objects with capability of ranking them and accessing underlying
references.