Update: Features & Improvements
This commit include a lot of changes that are not even compatible with
previous commit. Methods were renamed, modules refactored and improved
with also some other changes.

There as much changes within each module as they are contents. This commit
should make code more readable than previous commit and it contain more
detailed examples including even issues with the code.

This commit introduced DeepBlock class within block.py which allows easy
access to underlying item objects with nested block objects. Block objects
now updates priority of items when neccessary.

Priority of block has impact on items priorities and vice-verse. It also
gives control on how priority for block get calculated. Priority now is not just
number but priority can be anything that can be sorted like string, numbers, etc.

There are lot of changes that can take a while to explain. The issue with this  commit
is that some methods get called with initializer which has potential to cause problems
when the methods are overloaded. The dangers have been showed through comments in each
method affected.