Add, Update: highlevel.py,  __init__.py
New module 'highlevel.py' was added to source for handling high level functions
built on top of 'item' and 'block' modules. Function version of some methods
of block class were created in the module.

This is to allow immediate working with items without having to create block
object. There exists even functions for creating item and block instances
to avoid consumers from having to import item and block classes.

__init__.py was updated by importing 'highlevel.py' while exposing ites function
so that they can be accessed when import mimap(project name).