Fixes priorities updating bug
There was a bug that caused priories for items to be updated even
if they shouldnt be. Priorities for items shouldnt be uupdated if
priority for block is not provided.

Block priority which was influenced by items priorities should not
again influence items priorities. If that happens then its a bug
as 'x' cant dependent and be independent variable at same time. One
between 'x' and 'y' need to be independent or dependent not both.

The bug illustrates mathematical error which was not detected. This is
solved by only updating items priority only when priority for block is
provided. Priority for block should only be updated when its priority
is not provided.

That would ensure that either items or block influence another one not
both of them influencing each other. When there are no items and priority for
block is not provided then its better to raise exception.

Default priority should now be removed and it should no longer be 1. Each
item needs to be provided with priority else exception be raised if priority cannot
be found.