Update: source files
Priority is now optional but exception ill be raised if priority
is not provided when neccessary. If priority is not provided, it
is retrieved from object if not the exception get raised.

This applies to when creating item or block. This comes with limitation
of that None will be taken as if priority is not provided. That means
None wont be supported as priority as it already has another meaning.

Issues
------
- .get_item*() and find_item*() functions are returning None when item cannot
  be found. This would make it as if None is the item that was found but instead
  it should raise exception.

Progress
--------
- This is the final commit for release even if it contain issues.
- Neccessary features for first release have been added.
- The release is not production ready as there may be bugs.
- Thats because the project has no tests and it was not even tested manually.