Refactor & Improvements
Mimap library was planned to be library for associating object with priority.
But now there is a need for mimap like library but not associating objects
with priorities.

Similar classes to existing ones were added to try to make library more general.
The same classes were also used by existing classes which resulted in a lot
of changes without changing high level api.

As time goes on, priority features of mimap will be removed and only generic
classes remain. New library for priorities will be developed based on mimap
and it will also use the new mimap.