Added 'pemap' support
Support for pemap library was added even if it may pose problems. Pemap has
releasing breaking changing which mimap is currently using. Code that is available
on pemap was removed from mimap. Remeber that most of pemap code comes from mimap
but some of that code still exists in mimap.

Block could not use pemap due to lot of methods to be implemented. Its not easy as item
since block contain lot of method containing "*priority" while pemap contains "*value".
This release has dependency and wheather it works would depend on its dependency and how
it uses it.

Its bad idea that mimap inherit classes of pimap instead of compositing them. It could cause
issues if private attributes get changed or removed which would break the new classes.

If problems arise, it would be better to revert to version that does not depend on pemap.